requirements.txt
pandas
matplotlib
numpy
python-dotenv
requests
openpyxl
PyYAML
typer[all]